Fix TypeError in pixel transfer and ensure 16-bit PNG precision
- Explicitly set dtype=np.float32 when converting GPU buffer to numpy array to avoid 'incorrect sequence item type: d' error in foreach_set.
- Enabled float_buffer for PNG and TIFF images to preserve precision before saving at 16-bit.
- Maintained all previous improvements: color management via save_render(), native output paths, and UI cleanup.

+++ renderer.py
@@ -549,7 +549,8 @@ class Renderer:
 
         # In Blender 4.0+, buffer.to_list() followed by flattening is more reliable
         # to avoid shape mismatches or issues with foreach_set on direct gpu buffers
-        imageRes.pixels.foreach_set(np.array(buffer.to_list()).ravel())
+        # Explicitly use float32 to avoid "incorrect sequence item type: d" (double) error
+        imageRes.pixels.foreach_set(np.array(buffer.to_list(), dtype=np.float32).ravel())
 
         return imageRes
